bot.py
import os
from telegram import Update
from telegram.ext import Updater, CommandHandler, MessageHandler, Filters, CallbackContext

TOKEN_TELEGRAM = '5941419590:AAFZkLwttVADpoaMFloe6qg-FRLxKqS_MJs'

def start(update: Update, context: CallbackContext):
    update.message.reply_text("Hello, I'm your bot!")

def echo(update: Update, context: CallbackContext):
    update.message.reply_text(update.message.text)

def main():
    updater = Updater(TOKEN)
    dp = updater.dispatcher

    dp.add_handler(CommandHandler("start", start))
    dp.add_handler(MessageHandler(Filters.text & ~Filters.command, echo))

    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()


