requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
cursor==1.3.5
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.11.0
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.4
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-telegram-bot==20.2
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.9
sqlparse==0.4.3
stack-data==0.6.2
stevedore==5.0.0
terminado==0.17.1
tinycss2==1.2.1
torch==1.13.0a0+git49444c3
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.21.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
yarl==1.8.2
